use viz accessor, df.viz.histogram and df.x.viz.histogram
