+++ packages/vaex-viz/vaex/viz/mpl.py
@@ -1,7 +1,8 @@
 import ast
+import functools
 import numpy as np
 import logging
-from vaex.dataframe import DataFrame
+from vaex.dataframe import DataFrame, _hidden
 from vaex.utils import _parse_n, _parse_f, _ensure_string_from_expression, \
     _ensure_strings_from_expressions, _ensure_list,\
     _expand_limits, _expand_shape, _expand, _parse_reduction, _issequence
@@ -22,8 +23,15 @@ def add_plugin():
 def patch(f):
     '''Adds method f to the DataFrame class'''
     name = f.__name__
-    setattr(DataFrame, name, f)
+    setattr(DataFrame, name, _hidden(f))
     return f
+
+def viz_method(f):
+    @functools.wraps(f)
+    def wrapper(self, *args, **kwargs):
+        return f(self.df, *args, **kwargs)
+    from . import DataFrameAccessorViz as cls
+    setattr(cls, f.__name__, wrapper)
 
 
 patch(plot2d_vector)
@@ -34,6 +42,12 @@ patch(plot2d_contour)
 max_labels = 10
 
 @patch
+def plot1d(self, *args, **kwargs):
+    warnings.warn('`plot1d` is deprecated and it will be removed in version 5.x. Please use `df.viz.histogram` instead.')
+    self.viz.histogram(*args, **kwargs)
+
+
+@viz_method
 def histogram(self, x=None, what="count(*)", grid=None, shape=64, facet=None, limits=None, figsize=None, f="identity", n=None, normalize_axis=None,
               xlabel=None, ylabel=None, label=None,
               selection=None, show=False, tight_layout=True, hardcopy=None,
@@ -187,12 +201,12 @@ def histogram(self, x=None, what="count(*)", grid=None, shape=64, facet=None, li
 
 
 @patch
-def plot1d(self, x=None, **kwargs):
-    warnings.warn('`plot1d` is deprecated and it will be removed in version 5.x. Please `histogram` instead.')
-    histogram(self, x=x, **kwargs)
+def scatter(self, *args, **kwargs):
+    warnings.warn('`scatter` is deprecated and it will be removed in version 5.x. Please use `df.viz.scatter` instead.')
+    self.viz.scatter(*args, **kwargs)
 
 
-@patch
+@viz_method
 def scatter(self, x, y, xerr=None, yerr=None, cov=None, corr=None, s_expr=None, c_expr=None, labels=None, selection=None, length_limit=50000,
     length_check=True, label=None, xlabel=None, ylabel=None, errorbar_kwargs={}, ellipse_kwargs={}, **kwargs):
     """Viz (small amounts) of data in 2d using a scatter plot
@@ -291,10 +305,14 @@ def scatter(self, x, y, xerr=None, yerr=None, cov=None, corr=None, s_expr=None, 
             plt.errorbar(x_values, y_values, yerr=yerr_values, xerr=xerr_values, **errorbar_kwargs)
     return s
 
-# def plot(self, x=None, y=None, z=None, axes=[], row=None, agg=None, extra=["selection:none,default"], reduce=["colormap", "stack.fade"], f="log", n="normalize", naxis=None,
-
 
 @patch
+def plot(self, *args, **kwargs):
+    warnings.warn('`plot` is deprecated and it will be removed in version 5.x. Please `df.viz.heatmap` instead.')
+    self.viz.heatmap(*args, **kwargs)
+
+
+@viz_method
 def heatmap(self, x=None, y=None, z=None, what="count(*)", vwhat=None, reduce=["colormap"], f=None,
             normalize="normalize", normalize_axis="what",
             vmin=None, vmax=None,
@@ -859,6 +877,80 @@ def heatmap(self, x=None, y=None, z=None, what="count(*)", vwhat=None, reduce=["
 
 
 @patch
-def plot(self, x=None, y=None, z=None, **kwargs):
-    warnings.warn('`plot` is deprecated and it will be removed in version 5.x. Please `heatmap` instead.')
-    heatmap(self, x=x, y=y, z=z, **kwargs)
+def healpix_plot(self, *args, **kwargs):
+    warnings.warn('`healpix_plot` is deprecated and it will be removed in version 5.x. Please `df.viz.healpix_heatmap` instead.')
+    self.viz.healpix_heatmap(*args, **kwargs)
+
+
+@viz_method
+def healpix_heatmap(self, healpix_expression="source_id/34359738368", healpix_max_level=12, healpix_level=8, what="count(*)", selection=None,
+                    grid=None,
+                    healpix_input="equatorial", healpix_output="galactic", f=None,
+                    colormap="afmhot", grid_limits=None, image_size=800, nest=True,
+                    figsize=None, interactive=False, title="", smooth=None, show=False, colorbar=True,
+                    rotation=(0, 0, 0), **kwargs):
+    """Viz data in 2d using a healpix column.
+
+    :param healpix_expression: {healpix_max_level}
+    :param healpix_max_level: {healpix_max_level}
+    :param healpix_level: {healpix_level}
+    :param what: {what}
+    :param selection: {selection}
+    :param grid: {grid}
+    :param healpix_input: Specificy if the healpix index is in "equatorial", "galactic" or "ecliptic".
+    :param healpix_output: Plot in "equatorial", "galactic" or "ecliptic".
+    :param f: function to apply to the data
+    :param colormap: matplotlib colormap
+    :param grid_limits: Optional sequence [minvalue, maxvalue] that determine the min and max value that map to the colormap (values below and above these are clipped to the the min/max). (default is [min(f(grid)), max(f(grid)))
+    :param image_size: size for the image that healpy uses for rendering
+    :param nest: If the healpix data is in nested (True) or ring (False)
+    :param figsize: If given, modify the matplotlib figure size. Example (14,9)
+    :param interactive: (Experimental, uses healpy.mollzoom is True)
+    :param title: Title of figure
+    :param smooth: apply gaussian smoothing, in degrees
+    :param show: Call matplotlib's show (True) or not (False, defaut)
+    :param rotation: Rotatate the plot, in format (lon, lat, psi) such that (lon, lat) is the center, and rotate on the screen by angle psi. All angles are degrees.
+    :return:
+    """
+    # plot_level = healpix_level #healpix_max_level-reduce_level
+    import healpy as hp
+    import pylab as plt
+    if grid is None:
+        reduce_level = healpix_max_level - healpix_level
+        NSIDE = 2**healpix_level
+        nmax = hp.nside2npix(NSIDE)
+        # print nmax, np.sqrt(nmax)
+        scaling = 4**reduce_level
+        # print nmax
+        epsilon = 1. / scaling / 2
+        grid = self._stat(what=what, binby="%s/%s" % (healpix_expression, scaling), limits=[-epsilon, nmax - epsilon], shape=nmax, selection=selection)
+    if grid_limits:
+        grid_min, grid_max = grid_limits
+    else:
+        grid_min = grid_max = None
+    f_org = f
+    f = _parse_f(f)
+    if smooth:
+        if nest:
+            grid = hp.reorder(grid, inp="NEST", out="RING")
+            nest = False
+        # grid[np.isnan(grid)] = np.nanmean(grid)
+        grid = hp.smoothing(grid, sigma=np.radians(smooth))
+    fgrid = f(grid)
+    coord_map = dict(equatorial='C', galactic='G', ecliptic="E")
+    fig = plt.gcf()
+    if figsize is not None:
+        fig.set_size_inches(*figsize)
+    what_label = what
+    if f_org:
+        what_label = f_org + " " + what_label
+    f = hp.mollzoom if interactive else hp.mollview
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore")
+        coord = coord_map[healpix_input], coord_map[healpix_output]
+        if coord_map[healpix_input] == coord_map[healpix_output]:
+            coord = None
+        f(fgrid, unit=what_label, rot=rotation, nest=nest, title=title, coord=coord,
+            cmap=colormap, hold=True, xsize=image_size, min=grid_min, max=grid_max, cbar=colorbar, **kwargs)
+    if show:
+        plt.show()
